Adding typeit Tool
This tool helps to copy text from Clipboard and paste
it where ever you want. And since it emulates a
keyboard typing we don't need to worry where its being
written to. It works under **GUI only** rememeber that
while using it. Also it has 2 library dependencies
that need to be take care.